+++ src/ploomber/util/markup.py
@@ -4,12 +4,21 @@ from pygments.formatters import html
 import mistune
 
 
+try:
+    mistune_version = int(mistune.__version__[0])
+except ValueError:
+    mistune_version = 0
+
+
+mistune_recent = mistune_version >= 2
+
+
 def markdown_to_html(md):
     """
     Convert markdown to HTML with syntax highlighting, works with old and
     new versions of mistune
     """
-    if mistune.__version__[0] == "2":
+    if mistune_recent:
 
         class HighlightRenderer(mistune.HTMLRenderer):
             def block_code(self, code, lang=None):