+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.22.5dev
 
+* [Fix] Compatibility with mistune 3
+
 ## 0.22.4 (2023-06-01)
 
 * [Feature] Add `executor` option to NotebookRunner to choose ploomber-engine or papermill